Add rawdata.ts for hardcoded data retrieval
