+++ src/dnd/sources.ts
@@ -1,8 +1,6 @@
-import { readFileSync } from 'fs';
-import { join } from 'path';
-import * as vm from 'vm';
 import { Databank } from '../data';
 import { entrySort } from '../util';
+import { rawData } from '../5etools-conversion/rawdata';
 
 export interface ParsedSource {
     id: string;
@@ -12,38 +10,6 @@ export interface ParsedSource {
     published: string | null;
     author: string | null;
     group: string;
-}
-
-function loadParser(): void {
-    if ((globalThis as any).Parser) return;
-
-    const parserPath = join(__dirname, '..', '..', '5etools-src', 'js', 'parser.js');
-    const parserCode = readFileSync(parserPath, 'utf8');
-    const sandbox = {
-        globalThis: {} as any,
-        console,
-        setTimeout,
-        clearTimeout,
-        setInterval,
-        clearInterval,
-        Parser: undefined as any,
-    };
-
-    sandbox.globalThis = sandbox;
-    vm.runInNewContext(parserCode, sandbox, { filename: parserPath });
-
-    (globalThis as any).Parser = sandbox.Parser;
-}
-
-/**
- * Resolves the display name or abbreviation for a given source.
- * This function relies on the global `Parser` object from the **5e-tools** submodule.
- * If the source could not be resolved, this returns the backend source instead.
- */
-function getDisplayName(source: string): string {
-    loadParser();
-    const parser = (globalThis as any).Parser;
-    return parser?.SOURCE_JSON_TO_ABV?.[source] || source;
 }
 
 export function getSources(data: Databank): ParsedSource[] {
@@ -57,7 +23,7 @@ export function getSources(data: Databank): ParsedSource[] {
         sources.push({
             id: book.id,
             name: book.name,
-            displayName: getDisplayName(book.source),
+            displayName: rawData.getSourceDisplayName(book.source),
             source: book.source || book.id,
             published: book.published,
             author: book.author ?? null,
@@ -71,7 +37,7 @@ export function getSources(data: Databank): ParsedSource[] {
         sources.push({
             id: source,
             name: source,
-            displayName: getDisplayName(source),
+            displayName: rawData.getSourceDisplayName(source),
             source,
             published: null,
             author: null,
